Refactor models to eliminate cyclic imports
